+++ lib/build_info.dart
@@ -1,4 +1,4 @@
 final kBuildInfo = const String.fromEnvironment(
   'BUILD_INFO',
-  defaultValue: "You've found an easter egg! 🥚",
+  defaultValue: "You've found an Easter egg! 🥚",
 ).replaceAll(' | ', '\n');