+++ lib/link_constants.dart
@@ -1,0 +1,4 @@
+const repoLink = 'https://github.com/RealNoitasinagro/epoch';
+const changelogLink = '$repoLink/blob/main/CHANGELOG.md';
+const timezoneLink = 'https://pub.dev/packages/timezone';
+const ianaLink = 'https://www.iana.org/time-zones';